Rename IndexEEle.java lab1 problem2 to IndexEEle-> lab1 problem2.java
You are given a sorted (from smallest to largest) array A of n distinct integers which can be positive, negative, or zero. You want to decide whether or not there is an index i such that A[i] = i. Design the fastest algorithm that you can for solving this problem.